Anant Khandaka/requirements_voice_assistant.txt
autogen-agentchat
autogen-ext[ollama]
autogen-ext[chromadb]
autogen-core
python-dotenv
pyaudio
pygame
httpx
aiohttp
requests
aiofiles
fastapi
uvicorn
python-multipart
pydantic
deepgram-sdk
pdfplumber
python-docx